pyproject.toml
[project]
name = "mindmapagent"
version = "0.1.0"
description = "Collaborative brainstorming web app with LLM-powered mindmap generation"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.42.0",
    "flet>=0.25.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
]
